# elearningMobile/backend/api-contract.yaml
openapi: 3.0.1
info:
  title: ElearningMobile API (mock)
  version: '0.1.0'
paths:
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/signup:
    post:
      summary: Sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/lessons:
    get:
      summary: List lessons
      responses:
        '200':
          description: A list of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LessonSummary'
  /api/lessons/{id}:
    get:
      summary: Lesson detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonDetail'
        '404':
          description: Not found
  /api/assessments/submit:
    post:
      summary: Submit assessment answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentSubmission'
      responses:
        '200':
          description: Submission result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResult'
  /api/progress:
    get:
      summary: Get user progress
      responses:
        '200':
          description: User progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progress'
    post:
      summary: Save user progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Progress'
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true

components:
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
    SignupRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: mock-token
        user:
          type: object
          properties:
            id:
              type: string
              example: u1
            email:
              type: string
              example: user@example.com
    LessonSummary:
      type: object
      properties:
        id:
          type: string
          example: '1'
        title:
          type: string
          example: '1987 Constitution — Overview'
        excerpt:
          type: string
          example: 'Introduction to the 1987 Philippine Constitution'
    LessonDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        excerpt:
          type: string
        images:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            type: string
        content:
          type: string
    AssessmentSubmission:
      type: object
      properties:
        lessonId:
          type: string
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              answer:
                type: string
    AssessmentResult:
      type: object
      properties:
        score:
          type: number
          example: 85
        passed:
          type: boolean
          example: true
    Progress:
      type: object
      properties:
        dailyPoints:
          type: number
          example: 10
        lastLessonId:
          type: string
          example: '2'